disable default inits for ExpressibleByUnicodeScalarLiteral, ExpressibleByExtendedGraphemeClusterLiteral
The standard library contains the following definition:

```swift
extension ExpressibleByExtendedGraphemeClusterLiteral
  where ExtendedGraphemeClusterLiteralType == UnicodeScalarLiteralType {

  @_transparent
  public init(unicodeScalarLiteral value: ExtendedGraphemeClusterLiteralType) {
    self.init(extendedGraphemeClusterLiteral: value)
  }
}
```

If I declare a `NewType` with a conformance to
`ExpressibleByExtendedGraphemeClusterLiteral', like this:

```swift
@NewType<String> struct Name: ExtendedGraphemeClusterLiteralType { }
```

then I get an error because the default `init` provided by the
standard library conflicts with the default `init` I provide for
`ExpressibleByExtendedGraphemeClusterLiteral`. I couldn't find a way to
make my default `init` take precedence.

The standard library contains a similar extension for
`ExpressibleByStringLiteral` where StringLiteralType ==
ExtendedGraphemeClusterLiteralType` which conflicts with my default
`init` for `ExpressibleByExtendedGraphemeClusterLiteral`.

To work around the problem, I've disabled my default
`init`s for those protocols. I only provide a default
`init` for `ExpressibleByStringLiteral`. This covers
the most common case, where the `NewType` conforms to
`ExpressibleByStringLiteral`, but doesn't help for a `NewType`
that conforms only to `ExpressibleByUnicodeScalarLiteral` or
`ExpressibleByExtendedGraphemeClusterLiteral`.